add test for a constant T and V domain liquid reactor
simulation data comes from the RMG simulator

There is a slight ~1.5% difference in octane concentration between RMS and RMG at the chosen time.
I currently believe this is due to improvements in the calculation of diffusive rates in RMS not yet implemented in RMG since that difference
should give discrepencies of about this size.