Fix time index for fluxdiagrams
Previously, the filename used `t`, but `t` was the index for an array storing time points. Now, the filename uses `ts[t]` so it will save with the actual time point in seconds. 

The filename was also updated to include the radius, since previously running `plotfluxdiagrams` for the same time point with a different radius would overwrite the old file.